Cleaned up all warnings under VS2012

+++ remote_writer_memento.cpp
@@ -24,8 +24,8 @@ remote_writer_memento::remote_writer_memento(const configurator& cfg)
 {
     set_status_origin("remote_writer_memento");
     set_handler("host", [this] (const std::string& name) { host_ = name; });
-    set_handler("port", [this] (const std::string& port) { port_ = std::stoul(port); });
-    set_handler("unsent_cache_max", [this] (const std::string& sz) { unsent_cache_max_ = stoull(sz); });
+    set_handler("port", [this] (const std::string& port) { port_ = static_cast<std::uint16_t>(std::stoul(port)); });
+    set_handler("unsent_cache_max", [this] (const std::string& sz) { unsent_cache_max_ = static_cast<std::size_t>(stoull(sz)); });
 }
 
 }
